re go validation in service package
